--- src/Schema/Card/Card.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace Twigger\Blade\Schema\Card;
-
-use Twigger\Blade\Foundation\SchemaDefinition;
-
-abstract class Card extends SchemaDefinition
-{
-
-    public static function tag(): string
-    {
-        return 'card';
-    }
-
-    public static function defaultImplementation(): string
-    {
-        // TODO: Implement defaultImplementation() method.
-    }
-}
