+++ src/Themes/Bootstrap/Components/Card.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Twigger\Blade\Themes\Bootstrap\Components;
+
+class Card extends \Twigger\Blade\Schema\Card
+{
+
+    public function render()
+    {
+        return view('bootstrap-theme::card');
+    }
+}
