--- src/Themes/Bootstrap/views/card/card.blade.php
@@ -1,9 +1,0 @@
-<div class="card" {{ $attributes }}>
-    <div class="card-body">
-        <h5 class="card-title">{{ $title }}</h5>
-        <h6 class="card-subtitle mb-2 text-muted">{{  $subtitle }}</h6>
-        <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.</p>
-        <a href="#" class="card-link">Card link</a>
-        <a href="#" class="card-link">Another link</a>
-    </div>
-</div>
